examples/hello_world_opa/pyproject.toml
[project]
name = "hello-world-opa-example"
version = "0.1.0"
description = "Hello World Agent with OPA Integration Example"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.16",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "a2a-opa",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]  # Include all files in the current directory

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]